+++ sqlobject/dbconnection.py
@@ -1124,6 +1124,5 @@ from . import maxdb  # noqa
 from . import mssql  # noqa
 from . import mysql  # noqa
 from . import postgres  # noqa
-from . import rdbhost  # noqa
 from . import sqlite  # noqa
 from . import sybase  # noqa