+++ sqlobject/rdbhost/__init__.py
@@ -1,8 +1,0 @@
-from sqlobject.dbconnection import registerConnection
-
-
-def builder():
-    from . import rdbhostconnection
-    return rdbhostconnection.RdbhostConnection
-
-registerConnection(['rdbhost'], builder)